+++ R/status.R
@@ -34,53 +34,100 @@ renv_status_impl <- function(project, library, lockfile) {
   # check to see if we've initialized this project
   if (!renv_project_initialized(project)) {
     vwritef("* This project has not yet been initialized.")
-    return(FALSE)
+    return(list())
   }
 
-  # report missing lockfile
-  if (!file.exists(lockfile)) {
-    text <- if (identical(lockfile, renv_lockfile_path(project)))
-      "* This project has not yet been snapshotted -- 'renv.lock' does not exist."
-    else
-      sprintf("* Lockfile '%s' does not exist.", aliased_path(lockfile))
-    vwritef(text)
-  }
+  current  <- renv_status_check_missing_library(project, library)
+  recorded <- renv_status_check_missing_lockfile(project, lockfile)
 
-  # report missing library
-  projlib <- library[[1]]
-  if (!file.exists(projlib)) {
-    text <- if (identical(projlib, renv_paths_library(project = project)))
-      "* This project's private library is empty or does not exist."
-    else
-      sprintf("* Library '%s' is empty or does not exist.", aliased_path(projlib))
-    vwritef(text)
-  }
+  renv_status_check_used_packages(project, current)
+  renv_status_check_synchronized(project, recorded, current)
+  renv_status_check_cache(project)
 
-  # report status of cache
-  if (settings$use.cache(project = project))
-    renv_cache_diagnose()
-
-  # compare the lockfile with current state of library
-  curr <- local({
-    renv_scope_options(renv.verbose = FALSE)
-    snapshot(project = project, library = library, lockfile = NULL, force = TRUE)
-  })
-
-  lock <- renv_lockfile_load(project)
-  renv_status_report(lock, curr)
-  list(library = curr, lockfile = lock)
+  list(library = current, lockfile = recorded)
 
 }
 
-renv_status_report <- function(lock, curr) {
+renv_status_check_missing_lockfile <- function(project, lockfile) {
 
-  actions <- renv_lockfile_diff_packages(lock, curr)
-  if (empty(actions))
+  if (file.exists(lockfile))
+    return(renv_lockfile_read(lockfile))
+
+  if (identical(lockfile, renv_lockfile_path(project))) {
+    text <- "* This project has not yet been snapshotted -- 'renv.lock' does not exist."
+    vwritef(text)
+  } else {
+    fmt <- "* Lockfile %s does not exist."
+    vwritef(fmt, renv_path_pretty(lockfile))
+  }
+
+  list()
+
+}
+
+renv_status_check_missing_library <- function(project, library) {
+
+  projlib <- library[[1]]
+  if (file.exists(projlib)) {
+    renv_scope_options(renv.verbose = FALSE)
+    snapshotted <- snapshot(project = project,
+                            library = library,
+                            lockfile = NULL,
+                            force = TRUE)
+    return(snapshotted)
+  }
+
+  if (identical(projlib, renv_paths_library(project = project))) {
+    text <- "* This project's private library is empty or does not exist."
+    vwritef(text)
+  } else {
+    fmt <- "* Library %s is empty or does not exist."
+    vwritef(fmt, renv_paths_project(projlib))
+  }
+
+  list()
+
+}
+
+renv_status_check_used_packages <- function(project, current) {
+
+  db <- renv_installed_packages_base()
+
+  deps <- dependencies(project, quiet = TRUE)
+  used <- sort(unique(deps$Package))
+  records <- renv_records(current)
+
+  ignored <- c("R", db$Package, renv_project_ignored_packages(project), names(records))
+  missing <- setdiff(used, ignored)
+  if (empty(missing))
+    return(TRUE)
+
+  renv_pretty_print(
+    missing,
+    "The following package(s) are used in the project, but are not installed:",
+    c(
+      "Consider installing these packages, and then using `renv::snapshot()`",
+      "to record these packages in the lockfile."
+    ),
+    wrap = FALSE
+  )
+
+  FALSE
+
+}
+
+renv_status_check_synchronized <- function(project, recorded, current) {
+
+  # diff packages
+  actions <- renv_lockfile_diff_packages(recorded, current)
+  if (empty(actions)) {
     vwritef("* The project is already synchronized with the lockfile.")
+    return(TRUE)
+  }
 
   if ("install" %in% actions) {
     renv_pretty_print_records(
-      renv_records_select(curr, actions, "install"),
+      renv_records_select(current, actions, "install"),
       "The following package(s) are installed but not recorded in the lockfile:",
       "Use `renv::snapshot()` to add these packages to your lockfile."
     )
@@ -88,7 +135,7 @@ renv_status_report <- function(lock, curr) {
 
   if ("remove" %in% actions) {
     renv_pretty_print_records(
-      renv_records_select(lock, actions, "remove"),
+      renv_records_select(recorded, actions, "remove"),
       "The following package(s) are recorded in the lockfile but not installed:",
       "Use `renv::restore()` to install these packages."
     )
@@ -97,8 +144,8 @@ renv_status_report <- function(lock, curr) {
   rest <- c("upgrade", "downgrade", "crossgrade")
   if (any(rest %in% actions)) {
 
-    rlock <- renv_records(lock)
-    rcurr <- renv_records(curr)
+    rlock <- renv_records(recorded)
+    rcurr <- renv_records(current)
 
     matches <- actions[actions %in% rest]
     data <- data.frame(
@@ -118,5 +165,14 @@ renv_status_report <- function(lock, curr) {
     writeLines("Use `renv::restore()` to restore your library from the state of the lockfile.")
     writeLines("")
   }
+
+  FALSE
+
+}
+
+renv_status_check_cache <- function(project) {
+
+  if (settings$use.cache(project = project))
+    renv_cache_diagnose()
 
 }